+++ NEWS.md
@@ -6,7 +6,7 @@ Red Panda C++ Version 2.8
   - enhancement: "Run" / "Generate Assembly" for project source files
   - fix: Can't set project icon to "app.ico" in the project folder, if the project doesn't has icon.
   - fix: Resource compilation items is missing in the auto generated makefile, if the project's icon is removed and re-added.
-
+  - fix: Action "Run all problem cases" is triggered twice by one clicked.
 
 Red Panda C++ Version 2.7